Added types to the garbage collection phase.
Should implement actually giving pages back to the central free list. However,
that is a very sensitive part. When the object size of an object does not divide
a page, there can be overlap when pages are reused (even for the same object
type). Therefore, care must be taken when implementing this! First: write some
tests!